adding feature flag and Re-org folder structure
