+++ cypress.json
@@ -16,6 +16,7 @@
     "password": "admin",
     "ENDPOINT_WITH_PROXY": false,
     "MANAGED_SERVICE_ENDPOINT": false,
-    "VISBUILDER_ENABLED": false
+    "VISBUILDER_ENABLED": false,
+    "DATASOURCE_MANAGEMENT_ENABLED": false
   }
 }
